+++ ionq_core/api/default/clone_job.py
@@ -4,7 +4,7 @@
 
 from http import HTTPStatus
 from typing import Any, cast
-from urllib.parse import quote
+from ..._url import quote_path_param
 
 import httpx
 
@@ -33,7 +33,7 @@ def _get_kwargs(
 
     _kwargs: dict[str, Any] = {
         "method": "post",
-        "url": "/jobs/{uuid}/clone".format(uuid=quote(str(uuid), safe=""),),
+        "url": "/jobs/{uuid}/clone".format(uuid=quote_path_param(uuid),),
     }
 
     _kwargs["json"] = body.to_dict()