+++ ionq_core/api/default/get_job_probabilities.py
@@ -4,7 +4,7 @@
 
 from http import HTTPStatus
 from typing import Any, cast
-from urllib.parse import quote
+from ..._url import quote_path_param
 
 import httpx
 
@@ -38,7 +38,7 @@ def _get_kwargs(
 
     _kwargs: dict[str, Any] = {
         "method": "get",
-        "url": "/jobs/{uuid}/results/probabilities".format(uuid=quote(str(uuid), safe=""),),
+        "url": "/jobs/{uuid}/results/probabilities".format(uuid=quote_path_param(uuid),),
         "params": params,
     }